clafeLabs/SignalForge/requirements.txt
openai
yfinance
pandas
ta
requests
bs4
streamlit
plotly
python-dotenv